spx: made build histo idempotent
git-svn-id: http://svn.ilias.de/svn/ilias/branches/CaT/releases/4_4_seepex@58989 21b2c9ec-7c21-0410-8b45-9bfb8ed2bfc5

+++ Services/SPX/import/classes/class.spxBuildHisto.php
@@ -42,6 +42,9 @@ class spxBuildHisto {
 			
 			$participants = $crs_utils->getParticipants();
 			foreach ($participants as $participant) {
+				if ($crs_utils->getBookingStatusOf($participant) !== null) {
+					continue;
+				}
 				$crs_members_object->delete($participant);
 				$crs_utils->bookUser($participant);
